+++ packages/code/src/Code.tsx
@@ -12,6 +12,7 @@ import { Language, CodeProps, Mode } from './types';
 import Syntax from './Syntax';
 import Panel from './Panel';
 import WindowChrome from './WindowChrome';
+import { isComponentType } from '@leafygreen-ui/lib';
 
 export function hasMultipleLines(string: string): boolean {
   return string.trim().includes('\n');
@@ -153,6 +154,13 @@ function Code({
   highlightLines = [],
   languageOptions,
   onChange,
+  customActionButtons = [],
+  showCustomActionButtons = false,
+  usePortal = true,
+  portalClassName,
+  portalContainer,
+  scrollContainer,
+  popoverZIndex,
   ...rest
 }: CodeProps) {
   const scrollableElementRef = useRef<HTMLPreElement>(null);
@@ -161,6 +169,13 @@ function Code({
   const [showCopyBar, setShowCopyBar] = useState(false);
   const mode = darkMode ? Mode.Dark : Mode.Light;
   const isMultiline = useMemo(() => hasMultipleLines(children), [children]);
+
+  const filteredCustomActionIconButtons = customActionButtons.filter(
+    (item: React.ReactNode) => isComponentType(item, 'IconButton') === true,
+  );
+
+  const showCustomActionsInPanel =
+    showCustomActionButtons && !!filteredCustomActionIconButtons.length;
 
   const currentLanguage = languageOptions?.find(
     option => option.displayName === languageProp,
@@ -246,6 +261,18 @@ function Code({
     overflow: hidden;
   `;
 
+  const popoverProps = {
+    popoverZIndex,
+    ...(usePortal
+      ? {
+          usePortal,
+          portalClassName,
+          portalContainer,
+          scrollContainer,
+        }
+      : { usePortal }),
+  } as const;
+
   return (
     <div className={wrapperStyle}>
       {showWindowChrome && (
@@ -276,18 +303,22 @@ function Code({
 
         {/* Can make this a more robust check in the future */}
         {/* Right now the panel will only be rendered with copyable or a language switcher */}
-        {!showWindowChrome && (copyable || !!currentLanguage) && (
-          <Panel
-            language={currentLanguage}
-            languageOptions={languageOptions}
-            onChange={onChange}
-            contents={children}
-            onCopy={onCopy}
-            showCopyButton={showCopyBar}
-            darkMode={darkMode}
-            isMultiline={isMultiline}
-          />
-        )}
+        {!showWindowChrome &&
+          (copyable || !!currentLanguage || showCustomActionsInPanel) && (
+            <Panel
+              language={currentLanguage}
+              languageOptions={languageOptions}
+              onChange={onChange}
+              contents={children}
+              onCopy={onCopy}
+              showCopyButton={showCopyBar}
+              darkMode={darkMode}
+              isMultiline={isMultiline}
+              customActionButtons={filteredCustomActionIconButtons}
+              showCustomActionButtons={showCustomActionsInPanel}
+              {...popoverProps}
+            />
+          )}
       </div>
     </div>
   );