+++ packages/code/src/Panel.tsx
@@ -4,6 +4,7 @@ import { spacing } from '@leafygreen-ui/tokens';
 import CopyButton from './CopyButton';
 import LanguageSwitcher from './LanguageSwitcher';
 import { variantColors } from './globalStyles';
+import { PopoverProps } from './types';
 import {
   Mode,
   LanguageOption,
@@ -19,6 +20,13 @@ const copyStyle = css`
   flex-direction: column;
   flex-shrink: 0;
   padding-top: 6px;
+  align-items: center;
+  gap: ${spacing[1]}px;
+
+  svg {
+    width: 16px;
+    height: 16px;
+  }
 `;
 
 const singleLineCopyStyle = css`
@@ -68,6 +76,12 @@ function getPanelStyles(
     padding-right: 8px;
     background-color: ${colors[4]};
     border-bottom: 1px solid ${colors[1]};
+    gap: ${spacing[1]}px;
+
+    svg {
+      width: 16px;
+      height: 16px;
+    }
   `;
 }
 
@@ -78,7 +92,9 @@ type PanelProps = Partial<Omit<LanguageSwitcherProps, 'language'>> & {
   showCopyButton?: boolean;
   language?: LanguageOption;
   isMultiline?: boolean;
-};
+  customActionButtons?: Array<React.ReactNode>;
+  showCustomActionButtons?: boolean;
+} & PopoverProps;
 
 function Panel({
   language,
@@ -89,8 +105,23 @@ function Panel({
   showCopyButton,
   darkMode,
   isMultiline = false,
+  customActionButtons,
+  showCustomActionButtons,
+  usePortal,
+  portalClassName,
+  portalContainer,
+  scrollContainer,
+  popoverZIndex,
 }: PanelProps) {
   const mode = darkMode ? Mode.Dark : Mode.Light;
+
+  const popoverProps = {
+    popoverZIndex,
+    usePortal,
+    portalClassName,
+    portalContainer,
+    scrollContainer,
+  } as const;
 
   return (
     <div
@@ -105,6 +136,7 @@ function Panel({
             language={language}
             languageOptions={languageOptions}
             darkMode={darkMode}
+            {...popoverProps}
           />
         )}
 
@@ -115,6 +147,9 @@ function Panel({
           contents={contents}
           withLanguageSwitcher={!!language}
         />
+      )}
+      {showCustomActionButtons && (
+        <>{customActionButtons?.map((action: React.ReactNode) => action)}</>
       )}
     </div>
   );