+++ packages/text-area/src/TextArea.spec.tsx
@@ -1,5 +1,11 @@
 import React from 'react';
-import { render, fireEvent, screen } from '@testing-library/react';
+import {
+  render,
+  fireEvent,
+  screen,
+  getByLabelText,
+} from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import { axe } from 'jest-axe';
 import TextArea, { State } from './TextArea';
 
@@ -66,11 +72,62 @@ describe('packages/text-area', () => {
     });
   });
 
+  test('onBlur is invoked when focus leaves the textarea', () => {
+    const onBlur = jest.fn();
+
+    renderTextArea({ onBlur, ...defaultProps });
+
+    userEvent.tab(); // focus
+    userEvent.tab(); // blur
+
+    expect(onBlur).toHaveBeenCalledTimes(1);
+  });
+
   describe('when no label is supplied', () => {
     test('no label tag renders to the DOM', () => {
       renderTextArea();
 
       expect(screen.queryByRole('label')).not.toBeInTheDocument();
+    });
+  });
+
+  describe('validation callback', () => {
+    const setupValidationTest = () => {
+      const handleValidation = jest.fn();
+      const { container } = render(
+        <TextArea label="test" handleValidation={handleValidation} />,
+      );
+      const inputElement = getByLabelText(container, 'test');
+      return { handleValidation, inputElement };
+    };
+
+    test('is not called on focus', () => {
+      const { handleValidation } = setupValidationTest();
+      userEvent.tab(); // focus
+      expect(handleValidation).not.toHaveBeenCalled();
+    });
+
+    test('is not called on keypress before initial blur', () => {
+      const { handleValidation, inputElement } = setupValidationTest();
+      userEvent.tab(); // focus
+      userEvent.type(inputElement, `test`);
+      expect(handleValidation).not.toHaveBeenCalled();
+    });
+
+    test('is called on blur', () => {
+      const { handleValidation } = setupValidationTest();
+      userEvent.tab(); // focus
+      userEvent.tab(); // blur
+      expect(handleValidation).toHaveBeenCalledTimes(1);
+    });
+
+    test('is called on subsequent keypresses', () => {
+      const { handleValidation, inputElement } = setupValidationTest();
+      userEvent.tab(); // focus
+      userEvent.tab(); // blur
+      userEvent.tab(); // focus
+      userEvent.type(inputElement, `test`);
+      expect(handleValidation).toHaveBeenCalledTimes(5); // blur + keypress * 4
     });
   });