+++ packages/text-area/src/TextArea.tsx
@@ -5,7 +5,7 @@ import { css, cx } from '@leafygreen-ui/emotion';
 import InteractionRing from '@leafygreen-ui/interaction-ring';
 import { uiColors } from '@leafygreen-ui/palette';
 import { spacing, fontFamilies } from '@leafygreen-ui/tokens';
-import { useIdAllocator } from '@leafygreen-ui/hooks';
+import { useIdAllocator, useValidation } from '@leafygreen-ui/hooks';
 import { Description, Label } from '@leafygreen-ui/typography';
 
 export const State = {
@@ -128,6 +128,16 @@ type BaseTextAreaProps = HTMLElementProps<'textarea', HTMLTextAreaElement> & {
   description?: string;
   state?: State;
   errorMessage?: string;
+  handleValidation?: (value: string) => void;
+  /**
+   * Callback to be executed when the input stops being focused.
+   */
+  onBlur?: React.FocusEventHandler<HTMLTextAreaElement>;
+
+  /**
+   * Callback to be executed when the value of the input field changes.
+   */
+  onChange?: React.ChangeEventHandler<HTMLTextAreaElement>;
 };
 
 type AriaLabels = 'label' | 'aria-labelledby';
@@ -147,6 +157,8 @@ const TextArea: React.ComponentType<
     id: idProp,
     value: controlledValue,
     onChange,
+    onBlur,
+    handleValidation,
     'aria-labelledby': ariaLabelledby,
     ...rest
   }: TextAreaProps,
@@ -159,7 +171,18 @@ const TextArea: React.ComponentType<
   const [uncontrolledValue, setValue] = useState('');
   const value = isControlled ? controlledValue : uncontrolledValue;
 
-  const onValueChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+  // Validation
+  const validation = useValidation<HTMLTextAreaElement>(handleValidation);
+
+  const onBlurHandler: React.FocusEventHandler<HTMLTextAreaElement> = e => {
+    if (onBlur) {
+      onBlur(e);
+    }
+
+    validation.onBlur(e);
+  };
+
+  const onValueChange: React.ChangeEventHandler<HTMLTextAreaElement> = e => {
     if (onChange) {
       onChange(e);
     }
@@ -167,6 +190,8 @@ const TextArea: React.ComponentType<
     if (!isControlled) {
       setValue(e.target.value);
     }
+
+    validation.onChange(e);
   };
 
   if (!label && !ariaLabelledby) {
@@ -203,6 +228,7 @@ const TextArea: React.ComponentType<
           })}
           disabled={disabled}
           onChange={onValueChange}
+          onBlur={onBlurHandler}
           value={value}
         />
       </InteractionRing>