+++ packages/text-input/src/TextInput.tsx
@@ -7,7 +7,7 @@ import WarningIcon from '@leafygreen-ui/icon/dist/Warning';
 import InteractionRing from '@leafygreen-ui/interaction-ring';
 import { uiColors } from '@leafygreen-ui/palette';
 import { createDataProp, HTMLElementProps, Either } from '@leafygreen-ui/lib';
-import { useIdAllocator } from '@leafygreen-ui/hooks';
+import { useIdAllocator, useValidation } from '@leafygreen-ui/hooks';
 import { Description, Label } from '@leafygreen-ui/typography';
 
 const iconSelectorProp = createDataProp('icon-selector');
@@ -39,6 +39,19 @@ const Mode = {
 
 type Mode = typeof Mode[keyof typeof Mode];
 
+export const SizeVariant = {
+  XSmall: 'xsmall',
+  Small: 'small',
+  Default: 'default',
+  Large: 'large',
+} as const;
+
+export type SizeVariant = typeof SizeVariant[keyof typeof SizeVariant];
+
+export const BaseFontSize = 14 | 16;
+
+export type BaseFontSize = typeof BaseFontSize;
+
 interface TextInputProps extends HTMLElementProps<'input', HTMLInputElement> {
   /**
    * id associated with the TextInput component.
@@ -68,9 +81,14 @@ interface TextInputProps extends HTMLElementProps<'input', HTMLInputElement> {
   disabled?: boolean;
 
   /**
+   * Callback to be executed when the input stops being focused.
+   */
+  onBlur?: React.FocusEventHandler<HTMLInputElement>;
+
+  /**
    * Callback to be executed when the value of the input field changes.
    */
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onChange?: React.ChangeEventHandler<HTMLInputElement>;
 
   /**
    * The placeholder text shown in the input field before the user begins typing.
@@ -104,7 +122,21 @@ interface TextInputProps extends HTMLElementProps<'input', HTMLInputElement> {
 
   type?: TextInputType;
 
+  handleValidation?: (value: string) => void;
+
   ['aria-labelledby']?: string;
+
+  /**
+   *  determines the font size and padding.
+   */
+
+  sizeVariant?: SizeVariant;
+
+  /**
+   *  determines the base font size if sizeVariant is set to default.
+   */
+
+  baseFontSize?: BaseFontSize;
 }
 
 type AriaLabels = 'label' | 'aria-labelledby';
@@ -227,16 +259,26 @@ const interactionRingColor: Record<Mode, Record<'valid' | 'error', string>> = {
   },
 };
 
+interface SizeSet {
+  inputHeight: number;
+  inputText: number;
+  text: number;
+  lineHeight: number;
+  padding: number;
+}
+
 function getStatefulInputStyles({
   state,
   optional,
   mode,
   disabled,
+  sizeSet,
 }: {
   state: State;
   optional: boolean;
   mode: Mode;
   disabled: boolean;
+  sizeSet: SizeSet;
 }) {
   switch (state) {
     case State.Valid: {
@@ -262,11 +304,45 @@ function getStatefulInputStyles({
 
     default: {
       return css`
-        padding-right: ${optional ? 60 : 12}px;
+        padding-right: ${optional ? 60 : sizeSet.padding}px;
         border-color: ${colorSets[mode].defaultBorder};
       `;
     }
   }
+}
+
+function getSizeSets(baseFontSize: BaseFontSize, sizeVariant: SizeVariant) {
+  const sizeSets: Record<SizeVariant, SizeSet> = {
+    [SizeVariant.XSmall]: {
+      inputHeight: 22,
+      inputText: 12,
+      text: 14,
+      lineHeight: 20,
+      padding: 10,
+    },
+    [SizeVariant.Small]: {
+      inputHeight: 28,
+      inputText: 14,
+      text: 14,
+      lineHeight: 20,
+      padding: 10,
+    },
+    [SizeVariant.Default]: {
+      inputHeight: 36,
+      inputText: baseFontSize,
+      text: baseFontSize,
+      lineHeight: 20,
+      padding: 12,
+    },
+    [SizeVariant.Large]: {
+      inputHeight: 48,
+      inputText: 18,
+      text: 18,
+      lineHeight: 22,
+      padding: 16,
+    },
+  };
+  return sizeSets[sizeVariant];
 }
 
 /**
@@ -283,12 +359,14 @@ function getStatefulInputStyles({
  * @param props.optional Whether or not the field is optional.
  * @param props.disabled Whether or not the field is currently disabled.
  * @param props.onChange Callback to be executed when the value of the input field changes.
+ * @param props.onBlur Callback to be executed when the input stops being focused.
  * @param props.placeholder The placeholder text shown in the input field before the user begins typing.
  * @param props.errorMessage The message shown below the input field if the value is invalid.
  * @param props.state The current state of the TextInput. This can be none, valid, or error.
  * @param props.value The current value of the input field. If a value is passed to this prop, component will be controlled by consumer.
  * @param props.className className supplied to the TextInput container.
  * @param props.darkMode determines whether or not the component appears in dark mode.
+ * @param props.sizeVariant determines the size of the text and the height of the input.
  */
 const TextInput: React.ComponentType<
   React.PropsWithRef<AccessibleTextInputProps>
@@ -298,6 +376,7 @@ const TextInput: React.ComponentType<
       label,
       description,
       onChange,
+      onBlur,
       placeholder,
       errorMessage,
       optional = false,
@@ -308,7 +387,10 @@ const TextInput: React.ComponentType<
       value: controlledValue,
       className,
       darkMode = false,
+      sizeVariant = SizeVariant.Default,
       'aria-labelledby': ariaLabelledby,
+      handleValidation,
+      baseFontSize = 14,
       ...rest
     }: AccessibleTextInputProps,
     forwardRef: React.Ref<HTMLInputElement>,
@@ -318,8 +400,20 @@ const TextInput: React.ComponentType<
     const [uncontrolledValue, setValue] = useState('');
     const value = isControlled ? controlledValue : uncontrolledValue;
     const id = useIdAllocator({ prefix: 'textinput', id: propsId });
+    const sizeSet = getSizeSets(baseFontSize, sizeVariant);
 
-    function onValueChange(e: React.ChangeEvent<HTMLInputElement>) {
+    // Validation
+    const validation = useValidation<HTMLInputElement>(handleValidation);
+
+    const onBlurHandler: React.FocusEventHandler<HTMLInputElement> = e => {
+      if (onBlur) {
+        onBlur(e);
+      }
+
+      validation.onBlur(e);
+    };
+
+    const onValueChange: React.ChangeEventHandler<HTMLInputElement> = e => {
       if (onChange) {
         onChange(e);
       }
@@ -327,7 +421,9 @@ const TextInput: React.ComponentType<
       if (!isControlled) {
         setValue(e.target.value);
       }
-    }
+
+      validation.onChange(e);
+    };
 
     if (type !== 'search' && !label && !ariaLabelledby) {
       console.error(
@@ -348,12 +444,27 @@ const TextInput: React.ComponentType<
     return (
       <div className={cx(textInputStyle, className)}>
         {label && (
-          <Label darkMode={darkMode} htmlFor={id} disabled={disabled}>
+          <Label
+            darkMode={darkMode}
+            htmlFor={id}
+            disabled={disabled}
+            className={cx(css`
+              font-size: ${sizeSet.text}px;
+            `)}
+          >
             {label}
           </Label>
         )}
         {description && (
-          <Description darkMode={darkMode}>{description}</Description>
+          <Description
+            darkMode={darkMode}
+            className={cx(css`
+              font-size: ${sizeSet.text}px;
+              line-height: ${sizeSet.lineHeight}px;
+            `)}
+          >
+            {description}
+          </Description>
         )}
         <div className={inputContainerStyle}>
           <InteractionRing
@@ -378,6 +489,9 @@ const TextInput: React.ComponentType<
                 css`
                   color: ${colorSets[mode].inputColor};
                   background-color: ${colorSets[mode].inputBackgroundColor};
+                  font-size: ${sizeSet.inputText}px;
+                  height: ${sizeSet.inputHeight}px;
+                  padding-left: ${sizeSet.padding}px;
 
                   &:focus {
                     border: 1px solid ${colorSets[mode].inputBackgroundColor};
@@ -402,16 +516,24 @@ const TextInput: React.ComponentType<
                     }
                   }
                 `,
-                getStatefulInputStyles({ state, optional, mode, disabled }),
+                getStatefulInputStyles({
+                  state,
+                  optional,
+                  mode,
+                  disabled,
+                  sizeSet,
+                }),
               )}
               value={value}
               required={!optional}
               disabled={disabled}
               placeholder={placeholder}
               onChange={onValueChange}
+              onBlur={onBlurHandler}
               ref={forwardRef}
               id={id}
               autoComplete={disabled ? 'off' : rest?.autoComplete || 'on'}
+              aria-invalid={state === 'error'}
             />
           </InteractionRing>
 
@@ -452,6 +574,8 @@ const TextInput: React.ComponentType<
               errorMessageStyle,
               css`
                 color: ${colorSets[mode].errorMessage};
+                font-size: ${sizeSet.text}px;
+                line-height: ${sizeSet.lineHeight}px;
               `,
             )}
           >
@@ -477,6 +601,8 @@ TextInput.propTypes = {
   state: PropTypes.oneOf(Object.values(State)),
   value: PropTypes.string,
   className: PropTypes.string,
+  sizeVariant: PropTypes.oneOf(Object.values(SizeVariant)),
+  baseFontSize: PropTypes.oneOf([14, 16]),
 };
 
 export default TextInput;