+++ packages/code/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @leafygreen-ui/code
 
+## 9.4.0
+
+### Minor Changes
+
+- 548ca2c: Adds support for controlling the language switcher popover characteristics with `usePortal`, `portalClassName`, `portalContainer`, `scrollContainer`, and `popoverZIndex` props.
+- 548ca2c: Add support for Dart language
+
+## 9.3.0
+
+### Minor Changes
+
+- 70f3c2c: Add support for custom action buttons in the code component
+
+### Patch Changes
+
+- Updated dependencies [70f3c2c]
+- Updated dependencies [70f3c2c]
+  - @leafygreen-ui/hooks@7.1.1
+  - @leafygreen-ui/select@3.1.0
+
 ## 9.2.0
 
 ### Minor Changes