+++ packages/text-area/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @leafygreen-ui/text-area
 
+## 4.1.1
+
+### Patch Changes
+
+- 548ca2c: Restores the ability to pass onBlur handlers while using the useValidation hook.
+
+## 4.1.0
+
+### Minor Changes
+
+- d661688: Adds validation callback for TextArea and TextInput
+
+### Patch Changes
+
+- Updated dependencies [d661688]
+- Updated dependencies [d661688]
+  - @leafygreen-ui/lib@9.1.0
+  - @leafygreen-ui/hooks@7.1.0
+
 ## 4.0.3
 
 ### Patch Changes